requirements.txt
fastapi
uvicorn[standard]
jinja2
python-dotenv
pydantic-settings
sqlalchemy>=2.0
psycopg2-binary
slack_sdk>=3.0
bleach>=6.0
openai>=1.55.0
tzdata
